requirements.txt
python-dotenv~=1.0.1
langchain~=0.1.5
requests~=2.31.0
streamlit~=1.31.1
tweepy~=4.14.0